Add support for dynamically allocated stack
If xTaskCreate API is used to create a task, the task's stack is
allocated using pvPortMalloc. This places the task's stack in the
privileged data section (as all the heap memory is now placed in the
privileged data section).

We had a separate MPU region to grant a task access to its stack - If
the stack is in the privileged data section, this results in
overlapping MPU regions as privileged data section is already protected
using a separate MPU region. ARMv8-M does not allow overlapping MPU
regions and this results in a fault. This commit ensures to not use a
separate MPU region to protect the task's stack if it lies within the
privileged data section.

A side effect of this is that xTaskCreate API cannot be used to create
an unprivileged task as the task's stack will be in the privileged data
section and the task won't have access to it. xTaskCreateRestricted and
xTaskCreateRestrictedStatic API should be used to create unprivileged
tasks.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>