Fix: Two one line fixes, xTaskCheckForTimeOut() and ulTaskGenericNotifyValueClear().
# 72-character lines for description ###################################
ulTaskGenericNotifyValueClear() returned the notification value of the
currently running task, not the target task.  Now it returns the
notification value of the target task.

Some users expected xTaskCheckForTimeOut() to clear the 'last wake time'
value each time a timeout occurred, whereas it only did that in one path.
It now clears the last wake time in all paths that return that a timeout
occurred.