Better consistency checking in xPortSysTickHandler
* For Cortex-M ports it's crucial that the SysTick has the lowest
(logical) priority otherwise very strange things will happen when the
SysTick interrupts a critical section. Therefor a consistency check has
been added to the xPortSysTickHandler to make sure that a critical
section was not interrupted.

* Added comment in list.c to point in the right direction if the SysTick
has wrong interrupt priority. It's usually here you end up when
interrupting a critical section with the SysTick.