Support for optionally allocating stacks separate from the heap.
A new configuration macro (configSUPPORT_SEPARATE_STACK_ALLOCATION)
was added to allow stack allocation calls in tasks.c to be redirected
to a new portable layer allocation function.  This allows a client to
opt to add a separate stack allocator for dynamic stack allocation.
The default is the current behavior, which is to allocate stacks with
pvPortMalloc.