Send tmrCOMMAND_START_DONT_TRACE to front
When reloading an auto-reload timer, the timer task sends the
tmrCOMMAND_START_DONT_TRACE command if the new timer period has already
elapsed.

If the command queue already contains a stop or delete command for the
timer, then prior to this commit the tmrCOMMAND_START_DONT_TRACE command
is processed *after* the stop or delete command.  Consequently the timer
is restarted unexpectedly (in the stop case) or uses freed memory (in
the delete case).  After this commit, the tmrCOMMAND_START_DONT_TRACE
command is processed before the stop or delete command.