Do not use CMSIS names for interrupt handlers
We use the following 3 interrupt handlers in the port layer:
1. PendSV Handler
2. SVC Handler
3. SysTick Handler

Earlier we were using CMSIS names for these handlers which made it
difficult for the application writer to use an alternate handler with
the same CMSIS name. This commit changes the names of these handlers to
FreeRTOS specific ones, same as v7-M ports. The application writer
will need to map them to the CMSIS names, if needed, in their
FreeRTOSConfig.h:

 #define xPortPendSVHandler         PendSV_Handler
 #define vPortSVCHandler            SVC_Handler
 #define xPortSysTickHandler        SysTick_Handler

It was reported here - https://github.com/FreeRTOS/FreeRTOS-Kernel/issues/511

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>