Remove the need of compiler specific syntax
Introduce configKERNEL_PROVIDED_STATIC_MEMORY option which the
application can set to 1 to use the default implementations of
vApplicationGetIdleTaskMemory and vApplicationGetTimerTaskMemory
functions.

If an application enables static allocation (i.e. sets
configUSE_STATIC_ALLOCATION to 1) and does not provide the above 2
functions, it will result in linker error. The application has two
options:

1. Set configKERNEL_PROVIDED_STATIC_MEMORY to 1 to use the default
   implementations of these functions.
2. Provide implementations of these 2 functions.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>